stability: fold zram_linux.go into zram.go
Syncloud is always Linux, so the //go:build linux split served no
purpose. Move swaponFlags, the swapon/swapoff/loadModule methods, and
deviceSize into zram.go and drop the separate file.
